is_fav functionality
is_fav funcitonality working, prints this is a fav statement if the product is in the user fav list. Had to create functionality to get requests with a without token. Also fixed a few pages where class was called instead of className. Added message alert if user is not logged in but add to fav button is pressed.